requirements.txt
django==4.1.4
django-extensions
crispy-bootstrap5
django-formtools==2.3
django-jet-reboot
whitenoise
